Adding SystemInfoHelper - WIP
