+++ src/Elastic.Apm/AgentComponents.cs
@@ -2,6 +2,7 @@
 using System.IO;
 using Elastic.Apm.Api;
 using Elastic.Apm.Config;
+using Elastic.Apm.Helpers;
 using Elastic.Apm.Logging;
 using Elastic.Apm.Report;
 
@@ -20,38 +21,7 @@ namespace Elastic.Apm
 
 			Service = Service.GetDefaultService(ConfigurationReader);
 
-			var system = new Api.System();
-
-			try
-			{
-				if (File.Exists("/proc/self/cgroup"))
-				{
-					using (var sr = new StreamReader("/proc/self/cgroup"))
-					{
-						var line = sr.ReadLine();
-
-						while (line != null)
-						{
-							var fields = line.Split(':');
-							if (fields.Length == 3)
-							{
-								var dirAndId = fields[2].Split('/');
-
-								if (dirAndId.Length == 2)
-								{
-									var id = dirAndId[1].ToLower().EndsWith(".scope") ? dirAndId[1].Substring(".scope".Length) : dirAndId[1];
-									system.Container = new Container { Id = id };
-								}
-							}
-							line = sr.ReadLine();
-						}
-					}
-				}
-			}
-			catch (Exception e)
-			{
-				Logger.Error()?.LogException(e, "Failed reading container id");
-			}
+			var system = SystemInfoHelper.ReadContainerId(Logger);
 
 			PayloadSender = payloadSender ?? new PayloadSenderV2(Logger, ConfigurationReader, Service, system);
 			TracerInternal = new Tracer(Logger, Service, PayloadSender, ConfigurationReader);
